fix linting & circular dependency check
